adapt usage domains
to use new state method _output()
